use 64 bits unsigned integers in vars just like in meters.

+++ include/libopticonf/var.h
@@ -30,7 +30,7 @@ typedef enum vartype_e {
 /** Value union for var */
 typedef union varvalue_s {
     char            *sval; /**< String value (var owns memory) */
-    int              ival; /**< Integer value */
+    uint64_t         ival; /**< Integer value */
     double           dval; /**< Floating point value */
     varlink          arr; /**< Dictionary or array value */
 } varvalue;
@@ -58,31 +58,31 @@ void         var_free (var *);
 var         *var_find_key (var *, const char *);
 var         *var_find_index (var *, int);
 int          var_get_count (var *);
-int          var_get_int (var *);
+uint64_t     var_get_int (var *);
 double       var_get_double (var *);
 const char  *var_get_str (var *);
 var         *var_get_dict_forkey (var *, const char *);
 var         *var_get_array_forkey (var *, const char *);
-int          var_get_int_forkey (var *, const char *);
+uint64_t     var_get_int_forkey (var *, const char *);
 double       var_get_double_forkey (var *, const char *);
 const char  *var_get_str_forkey (var *, const char *);
 var         *var_get_dict_atindex (var *, int);
 var         *var_get_array_atindex (var *, int);
-int          var_get_int_atindex (var *, int);
+uint64_t     var_get_int_atindex (var *, int);
 double       var_get_double_atindex (var *, int);
 const char  *var_get_str_atindex (var *, int);
 
 void         var_new_generation (var *);
 void         var_clean_generation (var *);
-void         var_set_int (var *, int);
+void         var_set_int (var *, uint64_t);
 void         var_set_double (var *, double);
 void         var_set_str (var *, const char *);
-void         var_set_int_forkey (var *, const char *, int);
+void         var_set_int_forkey (var *, const char *, uint64_t);
 void         var_set_double_forkey (var *, const char *, double);
 void         var_set_str_forkey (var *, const char *, const char *);
 void         var_clear_array (var *);
 void         var_add_double (var *, double);
-void         var_add_int (var *, int);
+void         var_add_int (var *, uint64_t);
 void         var_add_str (var *, const char *);
 var         *var_add_array (var *);
 var         *var_add_dict (var *);
